+++ src/Concordium/Client/TransactionStatus.hs
@@ -10,6 +10,7 @@ import Concordium.Types.Execution
 import Concordium.Types
 
 import Concordium.Client.GRPC
+import Concordium.Client.GRPC2
 import Concordium.Client.Runner.Helper
 
 getSimpleTransactionStatus :: MonadIO m => TransactionHash -> ClientMonad m (Either String Value)
