+++ src/Concordium/Client/Types/TransactionStatus.hs
@@ -2,12 +2,14 @@
 {-# LANGUAGE OverloadedStrings #-}
 module Concordium.Client.Types.TransactionStatus where
 
+import Control.Monad.State (foldM)
 import qualified Data.Map.Strict as Map
 import Data.Aeson
 import Data.Aeson.TH
 
 import Concordium.Types.Execution
 import Concordium.Types
+import qualified Concordium.Types.Queries as Queries
 import Concordium.Utils
 
 data TransactionState = Received | Committed | Finalized | Absent deriving (Eq, Ord, Show)
@@ -24,6 +26,27 @@ data TransactionStatusResult' a = TransactionStatusResult
   deriving (Eq, Show)
 
 type TransactionStatusResult = TransactionStatusResult' ValidResult
+
+transactionStatusToTransactionStatusResult :: Queries.TransactionStatus -> Either String TransactionStatusResult
+transactionStatusToTransactionStatusResult tStatus = do
+  (tsrState, tsrResults) <- do
+    case tStatus of
+      Queries.Received ->
+        return (Received, Map.empty)
+      Queries.Committed bhToSummMap -> do
+        bhToSummlist <- foldM (\acc (bh, summM) -> do
+          case summM of
+            Nothing ->
+              err
+            Just summ ->
+              Right $ acc <> [(bh, summ)]) [] (Map.toList bhToSummMap)
+        return (Committed, Map.fromList bhToSummlist)
+      Queries.Finalized _ Nothing -> err
+      Queries.Finalized bHash (Just summ) ->
+        return (Finalized, Map.fromList [(bHash, summ)])
+  return TransactionStatusResult{..}
+  where
+    err = Left "transactionStatusToTransactionStatusResult: A transaction summary was missing."
 
 instance FromJSON a => FromJSON (TransactionStatusResult' a) where
   parseJSON Null = return TransactionStatusResult{tsrState = Absent, tsrResults = Map.empty}
